requirements.txt
dnspython==2.2.1
requests==2.26.0
tqdm==4.63.1
